Feat: Extract upload lecture logic into a hook and use web audio api for audio extraction
